Time: 16 ms (7.44%), Space: 16.5 MB (9.38%) - LeetHub

+++ 0229-majority-element-ii/0229-majority-element-ii.cpp
@@ -3,7 +3,7 @@ public:
     vector<int> majorityElement(vector<int>& nums) {
         vector<int> v;
         unordered_map<int, int> m;
-        int mini = ((int)nums.size() / 3)  + 1;
+        int mini = (nums.size() / 3)  + 1;
         for(int i : nums){
             m[i]++;
             if(m[i] == mini)
